requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
xgboost
imblearn
sentence-transformers
faiss-cpu
torch
ollama
streamlit
joblib